Refactor: Move getting Expenditure funding pot ID to a helper function
