+++ src/handlers/expenditures/helpers/index.ts
@@ -1,1 +1,2 @@
 export * from './getExpenditure';
+export * from './getExpenditureFundingPotId';
